Renamed rule for fill numbers to avoid recursion.
This bug was the bane of my existance.

Somehow due to metaclassing and the recursion this definition overwrote
the definition for all *number_sequence* functions even if they weren't
in the inheritance path. I discovered this bug by seeing that function
object (after much *HOW IS THIS WORKING?*) for number_sequence for both
CellParser and TallyParser are the same. TallyParser is not a subclass
of CellParser. The only common lineage is that they both are ultimately
subclasses of MCNP_Parser. This is some memory black magic nope.
I really need to get around to isssue #137.

Oh right this doesn't fix everything. It actually breaks almost
everything, BUT Tallies are now being parsed as intended.